--bind 0.0.0.0:$PORT
Here, "web" is the name of the process type, "app:app" refers to the Flask application named "app" in your project, and "--bind 0.0.0.0:$PORT" binds the application to the port specified by the environment variable "PORT". This allows Heroku to dynamically assign a port to the application when it is deployed.

To make the application accessible on web browsers, iOS, Android, tablets and mobile devices, you would need to develop the application with a responsive design that adapts to different screen sizes and resolutions. You may also need to develop native mobile applications for iOS and Android devices. The Procfile alone does not provide the functionality to make the application accessible on different platforms.