import the necessary libraries
In this function, we import the necessary libraries using import statements. We create a Flask app and define a simple route that returns "Hello, World!" when accessed. Finally, we use the if __name__ == '__main__': statement to ensure that the app only runs if the script is executed directly (as opposed to being imported as a module). We also set the debug flag to True so that we can see detailed error messages in case something goes wrong.

Note that this is just an example function, and you may need to modify it to suit your specific requirements. Additionally, you will need to make sure that you have installed the necessary packages using pip install Flask requests numpy in your virtual environment or system.