CSS
In the above CSS, we define the default styles for all devices first. Then we use media queries to adjust the styles for devices with a maximum width of 767px (phones) and a minimum width of 768px (tablets and desktops). This way, the styles will adapt to different screen sizes and provide a better user experience.